ai_learn/text2sql_postgres/requirements.txt
streamlit
pandas
psycopg2-binary
langchain-openai
python-dotenv
sqlalchemy
pg8000